requirements.txt
Flask
langchain
qdrant-client
langchain-nvidia-ai-endpoints
python-dotenv
langchain-community
gunicorn
fcntl